api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.5",
    "motor>=3.6.0",
    "pydantic-settings>=2.6.1",
    "pydantic>=2.9.2",
    "pyyaml>=6.0.2",
    "slowapi>=0.1.9",
    "tiktoken>=0.8.0",
    "curl-cffi>=0.7.3",
    "replicate>=1.0.3",
    "poe-api-wrapper>=1.7.0",
    "asyncstdlib>=3.13.0",
    "aiohttp>=3.11.7",
    "ujson>=5.10.0",
    "asgiref>=3.8.1",
    "gpytranslate>=1.5.1",
    "websockets>=14.1",
    "pillow>=11.1.0",
]
